glcs-to-glcp-migration-mainline/requirements.txt
aenum==3.1.11
aiohttp==3.8.4
aiosignal==1.3.1
allure-pytest==2.13.0
allure-python-commons==2.13.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
boto3==1.26.82
botocore==1.29.82
bs4==0.0.1
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
cleo==2.0.1
crashtest==0.4.1
cryptography==39.0.2
defusedxml==0.7.1
distlib==0.3.6
dulwich==0.21.3
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.92.0
filelock==3.9.0
flatdict==4.0.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
h11==0.14.0
#hpe-glcp-automation-lib @ file:///Users/kmamtora/Desktop/KARAN/QA/cumulus-testing/hpe_glcp_automation_lib-2.1.0-py3-none-any.whl
html5lib==1.1
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
installer==0.6.0
jaraco.classes==3.2.3
jira==2.0.0
jmespath==1.0.1
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
lxml==4.9.2
more-itertools==9.1.0
msgpack==1.0.4
multi-key-dict==2.0.3
multidict==6.0.4
names==0.3.0
oauth2client==4.1.3
oauthlib==3.2.2
okta==2.8.0
openapi-schema-validator==0.4.3
openapi-spec-validator==0.2.10
packaging==23.0
pbr==5.11.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
playwright==1.31.1
pluggy==1.0.0
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
prance==0.19.0
protobuf==4.22.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.5
pydash==6.0.2
pyee==9.0.4
PyJWT==2.6.0
pyotp==2.8.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.2.1
pytest-ordering==0.6
#pytest-pytestrail==0.10.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jenkins==1.7.0
python-jose==3.3.0
PyYAML==5.4.1
rapidfuzz==2.13.7
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
semver==2.13.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.25.0
testrail-api==1.12.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
trove-classifiers==2023.2.20
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.20.0
webdriver-manager==3.8.5
webencodings==0.5.1
xattr==0.10.1
xmltodict==0.13.0
yarl==1.8.2
zipp==3.15.0
